move gen_random to helpers
